# ackhandler: optimize received packet tracking