--- a/internal/ackhandler/received_packet_history_test.go
+++ b/internal/ackhandler/received_packet_history_test.go
@@ -22,23 +22,20 @@ var _ = Describe("receivedPacketHistory", func() {
 	Context("ranges", func() {
 		It("adds the first packet", func() {
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(1))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 4, End: 4}))
+			Expect(hist.ranges).To(Equal([]interval{{Start: 4, End: 4}}))
 		})
 
 		It("doesn't care about duplicate packets", func() {
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
 			Expect(hist.ReceivedPacket(4)).To(BeFalse())
-			Expect(hist.ranges.Len()).To(Equal(1))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 4, End: 4}))
+			Expect(hist.ranges).To(Equal([]interval{{Start: 4, End: 4}}))
 		})
 
 		It("adds a few consecutive packets", func() {
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
 			Expect(hist.ReceivedPacket(5)).To(BeTrue())
 			Expect(hist.ReceivedPacket(6)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(1))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 4, End: 6}))
+			Expect(hist.ranges).To(Equal([]interval{{Start: 4, End: 6}}))
 		})
 
 		It("doesn't care about a duplicate packet contained in an existing range", func() {
@@ -46,69 +43,71 @@ var _ = Describe("receivedPacketHistory", func() {
 			Expect(hist.ReceivedPacket(5)).To(BeTrue())
 			Expect(hist.ReceivedPacket(6)).To(BeTrue())
 			Expect(hist.ReceivedPacket(5)).To(BeFalse())
-			Expect(hist.ranges.Len()).To(Equal(1))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 4, End: 6}))
+			Expect(hist.ranges).To(Equal([]interval{{Start: 4, End: 6}}))
 		})
 
 		It("extends a range at the front", func() {
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
 			Expect(hist.ReceivedPacket(3)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(1))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 3, End: 4}))
+			Expect(hist.ranges).To(Equal([]interval{{Start: 3, End: 4}}))
 		})
 
 		It("creates a new range when a packet is lost", func() {
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
 			Expect(hist.ReceivedPacket(6)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(2))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 4, End: 4}))
-			Expect(hist.ranges.Back().Value).To(Equal(interval{Start: 6, End: 6}))
+			Expect(hist.ranges).To(Equal([]interval{
+				{Start: 4, End: 4},
+				{Start: 6, End: 6},
+			}))
 		})
 
 		It("creates a new range in between two ranges", func() {
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
 			Expect(hist.ReceivedPacket(10)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(2))
+			Expect(hist.ranges).To(HaveLen(2))
 			Expect(hist.ReceivedPacket(7)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(3))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 4, End: 4}))
-			Expect(hist.ranges.Front().Next().Value).To(Equal(interval{Start: 7, End: 7}))
-			Expect(hist.ranges.Back().Value).To(Equal(interval{Start: 10, End: 10}))
+			Expect(hist.ranges).To(Equal([]interval{
+				{Start: 4, End: 4},
+				{Start: 7, End: 7},
+				{Start: 10, End: 10},
+			}))
 		})
 
 		It("creates a new range before an existing range for a belated packet", func() {
 			Expect(hist.ReceivedPacket(6)).To(BeTrue())
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(2))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 4, End: 4}))
-			Expect(hist.ranges.Back().Value).To(Equal(interval{Start: 6, End: 6}))
+			Expect(hist.ranges).To(Equal([]interval{
+				{Start: 4, End: 4},
+				{Start: 6, End: 6},
+			}))
 		})
 
 		It("extends a previous range at the end", func() {
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
 			Expect(hist.ReceivedPacket(7)).To(BeTrue())
 			Expect(hist.ReceivedPacket(5)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(2))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 4, End: 5}))
-			Expect(hist.ranges.Back().Value).To(Equal(interval{Start: 7, End: 7}))
+			Expect(hist.ranges).To(Equal([]interval{
+				{Start: 4, End: 5},
+				{Start: 7, End: 7},
+			}))
 		})
 
 		It("extends a range at the front", func() {
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
 			Expect(hist.ReceivedPacket(7)).To(BeTrue())
 			Expect(hist.ReceivedPacket(6)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(2))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 4, End: 4}))
-			Expect(hist.ranges.Back().Value).To(Equal(interval{Start: 6, End: 7}))
+			Expect(hist.ranges).To(Equal([]interval{
+				{Start: 4, End: 4},
+				{Start: 6, End: 7},
+			}))
 		})
 
 		It("closes a range", func() {
 			Expect(hist.ReceivedPacket(6)).To(BeTrue())
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(2))
+			Expect(hist.ranges).To(HaveLen(2))
 			Expect(hist.ReceivedPacket(5)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(1))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 4, End: 6}))
+			Expect(hist.ranges).To(Equal([]interval{{Start: 4, End: 6}}))
 		})
 
 		It("closes a range in the middle", func() {
@@ -116,19 +115,20 @@ var _ = Describe("receivedPacketHistory", func() {
 			Expect(hist.ReceivedPacket(10)).To(BeTrue())
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
 			Expect(hist.ReceivedPacket(6)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(4))
+			Expect(hist.ranges).To(HaveLen(4))
 			Expect(hist.ReceivedPacket(5)).To(BeTrue())
-			Expect(hist.ranges.Len()).To(Equal(3))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 1, End: 1}))
-			Expect(hist.ranges.Front().Next().Value).To(Equal(interval{Start: 4, End: 6}))
-			Expect(hist.ranges.Back().Value).To(Equal(interval{Start: 10, End: 10}))
+			Expect(hist.ranges).To(Equal([]interval{
+				{Start: 1, End: 1},
+				{Start: 4, End: 6},
+				{Start: 10, End: 10},
+			}))
 		})
 	})
 
 	Context("deleting", func() {
 		It("does nothing when the history is empty", func() {
 			hist.DeleteBelow(5)
-			Expect(hist.ranges.Len()).To(BeZero())
+			Expect(hist.ranges).To(BeEmpty())
 		})
 
 		It("deletes a range", func() {
@@ -136,8 +136,7 @@ var _ = Describe("receivedPacketHistory", func() {
 			Expect(hist.ReceivedPacket(5)).To(BeTrue())
 			Expect(hist.ReceivedPacket(10)).To(BeTrue())
 			hist.DeleteBelow(6)
-			Expect(hist.ranges.Len()).To(Equal(1))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 10, End: 10}))
+			Expect(hist.ranges).To(Equal([]interval{{Start: 10, End: 10}}))
 		})
 
 		It("deletes multiple ranges", func() {
@@ -145,8 +144,7 @@ var _ = Describe("receivedPacketHistory", func() {
 			Expect(hist.ReceivedPacket(5)).To(BeTrue())
 			Expect(hist.ReceivedPacket(10)).To(BeTrue())
 			hist.DeleteBelow(8)
-			Expect(hist.ranges.Len()).To(Equal(1))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 10, End: 10}))
+			Expect(hist.ranges).To(Equal([]interval{{Start: 10, End: 10}}))
 		})
 
 		It("adjusts a range, if packets are delete from an existing range", func() {
@@ -156,8 +154,7 @@ var _ = Describe("receivedPacketHistory", func() {
 			Expect(hist.ReceivedPacket(6)).To(BeTrue())
 			Expect(hist.ReceivedPacket(7)).To(BeTrue())
 			hist.DeleteBelow(5)
-			Expect(hist.ranges.Len()).To(Equal(1))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 5, End: 7}))
+			Expect(hist.ranges).To(Equal([]interval{{Start: 5, End: 7}}))
 		})
 
 		It("adjusts a range, if only one packet remains in the range", func() {
@@ -165,16 +162,16 @@ var _ = Describe("receivedPacketHistory", func() {
 			Expect(hist.ReceivedPacket(5)).To(BeTrue())
 			Expect(hist.ReceivedPacket(10)).To(BeTrue())
 			hist.DeleteBelow(5)
-			Expect(hist.ranges.Len()).To(Equal(2))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 5, End: 5}))
-			Expect(hist.ranges.Back().Value).To(Equal(interval{Start: 10, End: 10}))
+			Expect(hist.ranges).To(Equal([]interval{
+				{Start: 5, End: 5},
+				{Start: 10, End: 10},
+			}))
 		})
 
 		It("keeps a one-packet range, if deleting up to the packet directly below", func() {
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
 			hist.DeleteBelow(4)
-			Expect(hist.ranges.Len()).To(Equal(1))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 4, End: 4}))
+			Expect(hist.ranges).To(Equal([]interval{{Start: 4, End: 4}}))
 		})
 
 		It("doesn't add delayed packets below deleted ranges", func() {
@@ -182,23 +179,21 @@ var _ = Describe("receivedPacketHistory", func() {
 			Expect(hist.ReceivedPacket(5)).To(BeTrue())
 			Expect(hist.ReceivedPacket(6)).To(BeTrue())
 			hist.DeleteBelow(5)
-			Expect(hist.ranges.Len()).To(Equal(1))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 5, End: 6}))
+			Expect(hist.ranges).To(Equal([]interval{{Start: 5, End: 6}}))
 			Expect(hist.ReceivedPacket(2)).To(BeFalse())
-			Expect(hist.ranges.Len()).To(Equal(1))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 5, End: 6}))
+			Expect(hist.ranges).To(Equal([]interval{{Start: 5, End: 6}}))
 		})
 
 		It("doesn't create more than MaxNumAckRanges ranges", func() {
 			for i := protocol.PacketNumber(0); i < protocol.MaxNumAckRanges; i++ {
 				Expect(hist.ReceivedPacket(2 * i)).To(BeTrue())
 			}
-			Expect(hist.ranges.Len()).To(Equal(protocol.MaxNumAckRanges))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 0, End: 0}))
+			Expect(hist.ranges).To(HaveLen(protocol.MaxNumAckRanges))
+			Expect(hist.ranges[0]).To(Equal(interval{Start: 0, End: 0}))
 			hist.ReceivedPacket(2*protocol.MaxNumAckRanges + 1000)
 			// check that the oldest ACK range was deleted
-			Expect(hist.ranges.Len()).To(Equal(protocol.MaxNumAckRanges))
-			Expect(hist.ranges.Front().Value).To(Equal(interval{Start: 2, End: 2}))
+			Expect(hist.ranges).To(HaveLen(protocol.MaxNumAckRanges))
+			Expect(hist.ranges[0]).To(Equal(interval{Start: 2, End: 2}))
 		})
 	})
 
@@ -211,17 +206,17 @@ var _ = Describe("receivedPacketHistory", func() {
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
 			Expect(hist.ReceivedPacket(5)).To(BeTrue())
 			ackRanges := hist.AppendAckRanges(nil)
-			Expect(ackRanges).To(HaveLen(1))
-			Expect(ackRanges[0]).To(Equal(wire.AckRange{Smallest: 4, Largest: 5}))
+			Expect(ackRanges).To(Equal([]wire.AckRange{{Smallest: 4, Largest: 5}}))
 		})
 
 		It("appends ACK ranges", func() {
 			Expect(hist.ReceivedPacket(4)).To(BeTrue())
 			Expect(hist.ReceivedPacket(5)).To(BeTrue())
 			ackRanges := hist.AppendAckRanges([]wire.AckRange{{Smallest: 1, Largest: 2}})
-			Expect(ackRanges).To(HaveLen(2))
-			Expect(ackRanges[0]).To(Equal(wire.AckRange{Smallest: 1, Largest: 2}))
-			Expect(ackRanges[1]).To(Equal(wire.AckRange{Smallest: 4, Largest: 5}))
+			Expect(ackRanges).To(Equal([]wire.AckRange{
+				{Smallest: 1, Largest: 2},
+				{Smallest: 4, Largest: 5},
+			}))
 		})
 
 		It("gets multiple ACK ranges", func() {
@@ -233,10 +228,11 @@ var _ = Describe("receivedPacketHistory", func() {
 			Expect(hist.ReceivedPacket(10)).To(BeTrue())
 			Expect(hist.ReceivedPacket(2)).To(BeTrue())
 			ackRanges := hist.AppendAckRanges(nil)
-			Expect(ackRanges).To(HaveLen(3))
-			Expect(ackRanges[0]).To(Equal(wire.AckRange{Smallest: 10, Largest: 11}))
-			Expect(ackRanges[1]).To(Equal(wire.AckRange{Smallest: 4, Largest: 6}))
-			Expect(ackRanges[2]).To(Equal(wire.AckRange{Smallest: 1, Largest: 2}))
+			Expect(ackRanges).To(Equal([]wire.AckRange{
+				{Smallest: 10, Largest: 11},
+				{Smallest: 4, Largest: 6},
+				{Smallest: 1, Largest: 2},
+			}))
 		})
 	})