And once again, using a simple slice (and a bunch of `copy`s) turns out to be significantly faster than the textbook data structure:
```
HistoryReceiveSequentialPackets-16            2.99ns ± 2%    2.50ns ± 2%  -16.64%  (p=0.008 n=5+5)
HistoryReceiveCommonCase-16                   3.25ns ± 2%    2.74ns ± 3%  -15.62%  (p=0.008 n=5+5)
HistoryReceiveSequentialPacketsWithGaps-16    12.5ns ± 1%     9.7ns ± 3%  -22.84%  (p=0.008 n=5+5)
HistoryReceiveReversePacketsWithGaps-16       48.3ns ± 2%    35.8ns ± 1%  -25.92%  (p=0.008 n=5+5)
HistoryIsDuplicate-16                         16.6ns ± 2%     8.3ns ± 1%  -49.80%  (p=0.008 n=5+5)
```