@marten-seemann 2024-07-19
Note to self: this code should make extensive use of the slices package. Ideally, we don’t copy elements around by hand at all.