@marten-seemann 2024-07-20
The `slices` package is quite neat! I was able to remove all instances of `copy` in [9d3d0a0](https://github.com/quic-go/quic-go/pull/4584/commits/9d3d0a076a957a7d15e35a3d5a4b28264ebfef6d).

However, it doesn't perform as well as simply using `copy`. It's still better than using a linked list though (and is even faster in what I - very crudely - declared to be the common case), so I don't feel very motivated to dig into this at this point.
```
name                                        old time/op    new time/op    delta
HistoryReceiveSequentialPackets-16            2.99ns ± 2%    2.18ns ± 1%  -27.06%  (p=0.008 n=5+5)
HistoryReceiveCommonCase-16                   3.25ns ± 2%    2.43ns ± 1%  -25.32%  (p=0.008 n=5+5)
HistoryReceiveSequentialPacketsWithGaps-16    12.5ns ± 1%    12.1ns ± 1%   -3.50%  (p=0.008 n=5+5)
HistoryReceiveReversePacketsWithGaps-16       48.3ns ± 2%    38.5ns ± 1%  -20.44%  (p=0.008 n=5+5)
HistoryIsDuplicate-16                         16.6ns ± 2%     7.9ns ± 1%  -52.44%  (p=0.008 n=5+5)
```