@sukunrt 2024-07-24
I should use `slices.Insert` more